[AIRFLOW-5498] Move Discovery hook to core
